+++ app/layouts/default.vue
@@ -1,7 +1,7 @@
 <template>
   <Header />
-  <div>
-    <slot />
+  <div class="w-full max-w-[1920px] mx-auto">
+      <slot></slot>
   </div>
   <Footer />
 </template>