+++ app/layouts/guest.vue
@@ -1,6 +1,6 @@
 <template>
-  <div>
-    <slot />
+  <div class="w-full max-w-[1920px] mx-auto">
+      <slot></slot>
   </div>
 </template>